Fix packaging of heuristics and move standalone modules into package
