+++ src/heuristics/__init__.py
@@ -1,0 +1,1 @@
+"""Heuristic engine and rules."""
