+++ src/pcap_tool/pipeline_app.py
@@ -14,7 +14,7 @@ from .enrich import service_guesser
 from .analyze import ErrorSummarizer, SecurityAuditor
 from .metrics_builder import MetricsBuilder
 from heuristics.engine import HeuristicEngine
-from llm_summarizer import LLMSummarizer
+from .llm_summarizer import LLMSummarizer
 from .utils import safe_int_or_default
 from .pipeline_helpers import (
     load_packets,
